refactor: split agents into separate modules
