+++ pa_core/agents/external_pa.py
@@ -1,0 +1,12 @@
+from __future__ import annotations
+from .types import Agent, Array
+
+class ExternalPAAgent(Agent):
+    def monthly_returns(self, r_beta: Array, alpha_stream: Array, financing: Array) -> Array:
+        if r_beta.shape != alpha_stream.shape or r_beta.shape != financing.shape:
+            raise ValueError("shape mismatch")
+        theta = float(self.extra.get("theta_extpa", 0.0))
+        return (
+            self.p.beta_share * (r_beta - financing)
+            + (self.p.beta_share * theta) * alpha_stream
+        )
